Moved AssertUtil to tests
